requirements.txt
gitingest
prowler
python-dotenv
pyyaml
requests
scoutsuite
selenium
tabulate
webdriver-manager
